+++ .storybook/main.ts
@@ -12,8 +12,6 @@ module.exports = {
   webpackFinal: async config => {
     config.resolve.alias['core-js/modules'] = '@storybook/core/node_modules/core-js/modules';
     config.resolve.alias['core-js/features'] = '@storybook/core/node_modules/core-js/features';
-    // mock the config file
-    config.resolve.alias['config'] = require.resolve('./config-mock.ts');
     // mock SWR
     config.resolve.alias['swr'] = require.resolve("./swr-mock.tsx");