+++ src/errors.ts
@@ -52,6 +52,25 @@ export class AppSetupError extends ApplicationError {
   }
 }
 
+/**
+ * Specifically for when fetch rejects, not for when you get an invalid response.
+ */
+export class FetchError extends ApplicationError {
+  url: string;
+
+  constructor(url: string, baseError?: Error) {
+    super(`${baseError ? baseError.message : ""}`);
+    Object.setPrototypeOf(this, FetchError.prototype);
+    this.name = "Fetch Error";
+    this.url = url;
+    this.info = {
+      title: "Fetch Error",
+      detail:
+        "The fetch promise for the requested resource was rejected. This is probably an offline, CORS, or other network error."
+    };
+  }
+}
+
 function isProblemDocument(
   details: OPDS1.ProblemDocument | OPDS1.AuthDocument
 ): details is OPDS1.ProblemDocument {
@@ -63,7 +82,7 @@ export class ServerError extends ApplicationError {
   info: OPDS1.ProblemDocument = {
     detail: "An unknown error server occurred.",
     title: "Server Error",
-    status: 418
+    status: 500
   };
   authDocument?: OPDS1.AuthDocument;
 
@@ -79,7 +98,7 @@ export class ServerError extends ApplicationError {
       // 401 errors return auth document instead of problem document
       // we will construct our own problem document.
       this.info = {
-        title: "No Authorized",
+        title: "Not Authorized",
         detail: "You are not authorized for the requested resource.",
         status: 401
       };
@@ -87,5 +106,6 @@ export class ServerError extends ApplicationError {
     } else if (isProblemDocument(details)) {
       this.info = details;
     }
+    this.name = `Server Error`;
   }
 }