+++ src/pages/_error.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 import ErrorComponent from "../components/Error";
 import { NextPage } from "next";
-import Bugsgnag from "utils/bugsnag";
 import { OPDS1 } from "interfaces";
+import track from "analytics/track";
 
 const Error: NextPage<{
   error?: OPDS1.ProblemDocument;
@@ -11,7 +11,7 @@ const Error: NextPage<{
 };
 
 Error.getInitialProps = ({ res, err }) => {
-  if (err) Bugsgnag.notify(err);
+  if (err) track.error(err);
   const statusCode = res?.statusCode ?? err?.statusCode ?? 404;
   return {
     error: {