+++ src/test-utils/fixtures/index.ts
@@ -8,3 +8,4 @@ export * from "./parsed-opds-feed";
 export * as opds2 from "./opds2";
 export * from "./collection";
 export * from "./user";
+export * from "./config";